test-fixtures/python-django/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "acme-api"
version = "1.0.0"
description = "Acme Corp Django REST API"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<5.2",
    "djangorestframework>=3.15,<4.0",
    "psycopg2-binary>=2.9.9",
    "celery[redis]>=5.4,<6.0",
    "django-cors-headers>=4.4,<5.0",
    "django-filter>=24.3",
    "gunicorn>=23.0",
    "pyjwt>=2.9",
    "python-dotenv>=1.0",
    "whitenoise>=6.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "bandit>=1.8",
    "safety>=3.2",
    "factory-boy>=3.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=app --cov-report=term-missing --cov-branch"

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DJ"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main"]
